sql: move source-related planning impls to their own module
