+++ src/sql/src/plan.rs
@@ -96,9 +96,9 @@ pub use notice::PlanNotice;
 pub use query::{ExprContext, QueryContext, QueryLifetime};
 pub use scope::Scope;
 pub use side_effecting_func::SideEffectingFunc;
+pub use statement::ddl::source::PgConfigOptionExtracted;
 pub use statement::ddl::{
-    AlterSourceAddSubsourceOptionExtracted, PgConfigOptionExtracted, PlannedAlterRoleOption,
-    PlannedRoleVariable,
+    AlterSourceAddSubsourceOptionExtracted, PlannedAlterRoleOption, PlannedRoleVariable,
 };
 pub use statement::{
     describe, plan, plan_copy_from, resolve_cluster_for_materialized_view, StatementContext,
